Some minor cleanup
Reorder methods on ResolvedVariantResult so attributes and capabilities are next to each other.
Remove CachingDependencyResultFactory. It may have been useful in the past, but it currently is not reused and is ineffective as caching.
Remove AbstractDependencyResult in favor of implementing getters manually
Add module version ID-accepting overload of DefaultModuleComponentSelector.newSelector for ease in testing