Complete ProjectMetadataController
This is the mechanism that configuration cache uses to serialize the dependency management state of a project during
project isolation. The existing serialization scheme was quite incomplete and did not serialize the entire set
of project state data. This commit enhances ProjectMetadataController to serialize all component and variant state,
thus bringing the serialized data to parity with the original data.