Create separate graph node for edge with requested artifacts
When an edge requests a specific artifact or artifacts from a node, create an adhoc node based
off of the original node, and use that in the graph -- instead of using the original node and then
handling the requested artifacts during artifact selection time.

This moves more toward a place where each node in a given graph always contributes the same artifacts
regardless of the edge pointing to it.