Introduce BuildLocations as replacement for BuildLayout
The internal BuildLayout service is not in an explicit
internal package despite not being part of the public API.
It is often mistakenly injected by users, and we even had
an userguide snippet using it. The confusing partly comes
from the fact that we have a public BuildLayout service,
which lives in another package.

Introducing a new internal service makes the distinction clear.